requirements.txt
Django==4.1
django-environ==0.9.0
djangorestframework==3.13.1
markdown==3.4.1
django-filter==22.1
django-cors-headers==3.13.0
mixer==7.2.2
